+++ frontend/src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
 
-import { DefaultLayoutComponent } from "./containers";
+import { DefaultLayoutComponent } from "./shared/containers";
 import { Page404Component } from "./views/pages/page404/page404.component";
 import { Page500Component } from "./views/pages/page500/page500.component";
 import { LoginComponent } from "./views/pages/login/login.component";
