+++ frontend/src/app/shared/containers/default-layout/default-footer/default-footer.component.spec.ts
