+++ frontend/src/app/shared/containers/default-layout/default-layout.component.ts
